pyproject.toml
[build-system]
requires = ["setuptools", "setuptools_scm"]
build-backend = "setuptools.build_meta"

[project]
name = "realtimeobserver"
description = "Monitoring of realtime data presence in passenger information systems"
authors = [{ name = "Sebastian Knopf", email = "mail@sebastian-knopf.de" }]
readme = "README.md"
license = {text = "Apache License (2.0)"}
dependencies = [
    "click",
    "lxml",
    "pytz",
    "pyyaml",
    "requests",
    "sqlobject"
]
requires-python = ">=3.10"

dynamic = ["version"]

[tool.setuptools_scm]
write_to = "src/realtimeobserver/version.py"